fix(examples): resolve sign conversion warning in body_inspector_example
Fix implicit sign conversion error when accessing vector with int index:
- Cast selectedBodyIndex to size_t when used as vector subscript
- Resolves -Werror,-Wsign-conversion on Clang and GCC

This ensures the example compiles on all CI platforms with strict
warning settings.

+++ examples/body_inspector_example.cpp
@@ -254,7 +254,7 @@ int main() {
 
         // Render body inspector for selected body
         if (selectedBodyIndex >= 0 && selectedBodyIndex < static_cast<int>(bodies.size())) {
-            auto& selectedBody = bodies[selectedBodyIndex];
+            auto& selectedBody = bodies[static_cast<size_t>(selectedBodyIndex)];
 
             // Create window title with body name
             char windowTitle[256];
